updated password generation and saving of changes
