+++ .gitignore
@@ -2,4 +2,5 @@
 **/__pycache__/**
 oldPassword
 newPassword
-hosts
+hosts_*
+**/hosts
