+++ diceword_lists/diceware_german.txt
