+++ diceword_lists/diceware_wordlist.txt
