+++ hosts
@@ -1,2 +1,8 @@
-evelyn:10.10.0.11:Password
-evelyn:10.10.0.12:Password
+evelyn,10.10.0.11
+evelyn,10.10.0.12
+evelyn,10.10.0.13
+evelyn,10.10.0.14
+evelyn,10.10.0.15
+evelyn,10.10.0.16
+evelyn,10.10.0.17
+evelyn,10.10.0.18
