+++ main.py
@@ -6,144 +6,74 @@
 
 # --- / 
 # -- / internal imports 
-
+from file_handling import filename_set_timestamp, readFromFile,gatherFilePath
+from dice_generation import generatePassword
 
 # --- / 
 # -- / external imports
 from typing import Optional
-from selenium import webdriver
-# from selenium.webdriver.firefox.options import Options
-from selenium.webdriver.common.by import By
 import time 
 import glob
 import datetime
 import os
 
-
-# --- /
-# -- / 
-
-def requestPassword(length:int) -> str | None:
-    '''
-    queries diceware.dmuth.org and retrieves a **password** consisting of **length** words
-    returns password if successful -> type **string**
-    else None at failure 
-    '''
-    # TODO better way to wait for result
-    # executing search for item:
-    try:
-        # TODO Extract to loadWebsite()
-        requestUrl = "https://diceware.dmuth.org/?debug={}&skip_animation".format(length)
-        browser = webdriver.Firefox()
-        browser.get(requestUrl)
-        time.sleep(4)
-    
-        passwordFields = browser.find_element(By.CLASS_NAME, "results_phrase_value")
-        extractedPassword = passwordFields.get_attribute("innerText")
-        browser.quit()
-        return extractedPassword
-    except: 
-        return None
-    
-# --- /
-# -- / 
-
-def writeNewPassword(passwordLength:int) -> bool:
-    '''
-    acquires new password from diceware website and saves it to **newPassword** 
-    returns **True** if successfull
-    '''
-    newPassword:str | None = requestPassword(passwordLength)
-    if (newPassword == None):
-        raise Exception("Password could not be obtained")
-    writeToFile("newPassword",newPassword)
-    return True
-    
-# --- /
-# -- / 
-def gatherFilePath(filename,recursiveSearch:bool=True) -> Optional[str]:
+def generate_password_for_each_device_in_list(host_file:str,length:int) -> bool:
     ''' 
-    function searching for **filename** in current directory
-    searches **recursively** relative from root of **environment** -> finds files in a directory too.
-    '''
-    convertedQueryPath = "**/"+filename
-    try:
-        listOfFiles = glob.glob(convertedQueryPath,recursive=recursiveSearch)
-        return listOfFiles[0];
-    except: 
-        return None 
-
-# --- /
-# -- / 
-
-def readFromFile(filename:str) -> Optional[str]:
-    '''
-    takes a filepath as argument -> type **string**
-    returns content of file or **None** if not found 
-    '''
-    filePath: Optional[str] = gatherFilePath(filename)
-    if (filePath == None):
-        return None
-    # no error checking necessary, is checked with aboves if condition
-    with open(filePath,"r") as foundFile:
-        fileContent:str = foundFile.read()
-    return fileContent
-
-# --- /
-# -- / 
-
-def writeToFile(filename:str,input:str) -> bool :
-    '''
-    inserts given **input** to the given filename
-    returns None if an error occured
-    '''
-    filePath: Optional[str] = gatherFilePath(filename)
-    if (filePath == None):
-        # no previous file was found:
-        filePath = filename
-    try:
-        createBackup(filePath)
-        with open(filePath,"w") as file:
-            file.write(input)
-        return True
-    except Exception as exception:
-        raise exception
-        
-# --- /
-# -- / 
-
-def createBackup(fileName):
-    '''
-    takes **filePath** and creates a new file containing this files content. 
-    the new file is constructed as follows: **filePath**+ **date_backup**
-    in case of errors, it returns **False**
-    **True** otherwise
-    '''
-    currentDate:str = str(datetime.datetime.now())
-    # suffix:str = ".backup"
-    backupName:str = "oldPassword"
-    fileContent:str | None = readFromFile(fileName)
-    if (fileContent == None):
-        return False
-    newFileName:str = "{}_{}".format(backupName,currentDate)
-    print("creating backup of file {}. Saving to --> {}".format(fileName,newFileName))
-    writeToFile(newFileName,fileContent)
+    @param1 string, denotes file to generate passwords for 
+    @param2 integer, denotes length of passwords ( amount of words per password)
+    function checks whether file exists, ( raises exception if not!) and
+    - reads every line 
+    - generates a password per line 
+    - writes the updated line to a new file, denoted with **param1**+timestamp
+    @returns bool, **True** if successfull, **False** if error occurred
     
-    return True
-
-# --- /
+    '''
+    
+    file_path:str | None = gatherFilePath(host_file)
+    if file_path == None:
+        raise Exception("host_file was not found, aborting")
+    
+    # going through file
+    #TODO Refactor 
+    updated_lines:list[str] = []
+    with open(file_path,"r",encoding="UTF-8") as hosts_file:
+        for line in hosts_file:
+            line =line.rstrip()
+            password_for_line:str = generatePassword(length) 
+            new_line = "{},{}".format(line,password_for_line)
+            updated_lines.append(new_line)
+            print( "updated password {}".format(new_line))
+    
+    # creating new file to update content
+    new_file_path:str = filename_set_timestamp(file_path)
+    try:
+        write_list_to_file(new_file_path,updated_lines)
+        return True
+    except:
+        return False
+    
+# --- / 
 # -- / 
 
-def gatherUserInput(displayedMessage:str) -> int:
-    #TODO check range (min,max)
+def write_list_to_file(filename:str,list_of_items:list[str])-> bool:
     '''
-    function that is checking for an according  
+    @param1 string, denoting filename to write to 
+    @param2 list[str], denoting which list to write to file
+    function truncates and generates file and inserts every item of the list in **a newline**.
+    @returns bool, **True** if successfull, **False** if errors occurred
+     
     '''
-    gatheredInput = input("{} \n".format(displayedMessage)) 
+    #TODO add backup-check
     try:
-        return int(gatheredInput)
-    except:
-        return gatherUserInput(displayedMessage)
+        with open(filename,"w",encoding='UTF-8') as new_file:
+            for item in list_of_items:
+                new_file.write("{}\n".format(item))
+        return True
+    except  Exception as error:
+        raise error
+        
+     
+    
 
 def interface():
     '''
@@ -164,14 +94,9 @@ def interface():
         clearScreen()
         print(programmHeader)
         displayMenu(menuOptions)
-        selection:int = gatherUserInput("enter an integer")
+        selection:int = int(input("enter an integer"))
         print("selected {} -> {}".format(selection,menuOptions[selection] ))
-        
         break
-
-        
-        
-        
         
 def clearScreen():
     # taken from https://stackoverflow.com/questions/2084508/clear-terminal-in-python#2084628
@@ -186,9 +111,6 @@ def displayMenu(options):
 # -- / 
 if __name__ == "__main__":
     print("executing script")
-    print(writeNewPassword(5))
-    # interface()
-    
-    # writeToFile("newPassword","Das sollte funktionieren")
+    generate_password_for_each_device_in_list("hosts",5)
